# [SPARK-27398][SQL] Refactoring of CreateJacksonParser.getStreamDecoder