## What changes were proposed in this pull request?

In the PR, I propose to simplify the `CreateJacksonParser.getStreamDecoder` method, and use `InputStreamReader` instead of creating `ReadableByteChannel` and do the channel decoding via `StreamDecoder.forDecoder`. Also I replaced returned type by `Reader` and renamed the method from `getStreamDecoder` to `getReader`.

## How was this patch tested?

by `JsonSuite`, `JsonFunctionsSuite` and `JsonExpressionsSuite`.